Rearrange Main Layout and Showcase Layout
Move blog carousel into Home page so it's only visible on the home page. No dedicated page for the carousel. ShowcaseLayout may not be necessary.

Minor updates to the main layout to make UI changes simpler in the future.

TODO add fade in and out on start and end edges of carousel. Make the padding for carousel responsive.